+++ src/actions/build_node_server.ts
@@ -5,6 +5,7 @@ import * as path from "path";
 import { install_declared_deps } from "../components/project_deps/install_declared_deps";
 import { ordering_project_deps } from "../components/project_deps/ordering_project_deps";
 import { sync_with_linked_modules } from "../components/project_deps/sync_with_linked_modules";
+import { envConfig } from "../config/env";
 import { prettierConfig } from "../config/prettier";
 import { tsLintConfig } from "../config/tslint";
 import { typescriptConfig } from "../config/typescript";
@@ -20,21 +21,21 @@ async function build_node_server() {
   prettierConfig.init(true);
 
   // Meta files init
-  dockerIgnore.init(true);
-  gitIgnore.init(true);
-  editorConfig.init(true);
-  npmrc.init(true);
+  dockerIgnore.init(envConfig, true);
+  gitIgnore.init(envConfig, true);
+  editorConfig.init(envConfig, true);
+  npmrc.init(envConfig, true);
 
   console.log("");
 
-  await install_declared_deps();
-  await ordering_project_deps();
-  await sync_with_linked_modules();
+  await install_declared_deps(envConfig);
+  await ordering_project_deps(envConfig);
+  await sync_with_linked_modules(envConfig);
 
   console.log(chalk.bold.blue(`[ BUILD_NODE_SERVER ][ START ]`));
-  const startTime = moment();
   console.log("");
 
+  const startTime = moment();
   const result = spawn.sync(
     "tsc",
     [
@@ -54,6 +55,7 @@ async function build_node_server() {
     },
   );
 
+  // ! Обработка сигнала процесса
   if (result.signal) {
     if (result.signal === "SIGKILL") {
       console.log(