+++ src/actions/sync.ts
@@ -4,6 +4,7 @@ import { sync_with_linked_modules } from "../components/project_deps/sync_with_l
 import { watch_deps } from "../components/watch_deps";
 import { envConfig } from "../config/env";
 import { prettierConfig } from "../config/prettier";
+import { rearguardConfig } from "../config/rearguard";
 import { tsLintConfig } from "../config/tslint";
 import { typescriptConfig } from "../config/typescript";
 import { dockerIgnore } from "../meta/dockerignore";
@@ -18,19 +19,19 @@ async function sync() {
   prettierConfig.init(true);
 
   // Meta files init
-  dockerIgnore.init(true);
-  gitIgnore.init(true);
-  editorConfig.init(true);
-  npmrc.init(true);
+  dockerIgnore.init(envConfig, true);
+  gitIgnore.init(envConfig, true);
+  editorConfig.init(envConfig, true);
+  npmrc.init(envConfig, true);
 
   console.log("");
 
-  await install_declared_deps();
-  await ordering_project_deps();
-  await sync_with_linked_modules();
+  await install_declared_deps(envConfig);
+  await ordering_project_deps(envConfig);
+  await sync_with_linked_modules(envConfig);
 
   if (envConfig.isWatch) {
-    watch_deps();
+    watch_deps(envConfig, rearguardConfig);
   }
 }