+++ src/meta/Typings/Typings.ts
@@ -1,7 +1,3 @@
-import chalk from "chalk";
-import * as fs from "fs";
-import * as path from "path";
-import { envConfig } from "../../config/env";
 import { IMetaFile } from "../../interfaces/metaFile/IMetaFile";
 import { MetaFile } from "../MetaFile";