+++ src/meta/monorepo_gitignore/MonorepoGitIgnore.ts
@@ -1,7 +1,7 @@
 import chalk from "chalk";
 import * as fs from "fs";
 import * as path from "path";
-import { envConfig } from "../../config/env";
+import { IEnvConfig } from "../../interfaces/config/IEnvConfig";
 import { IMetaFile } from "../../interfaces/metaFile/IMetaFile";
 
 export class MonorepoGitIgnore implements IMetaFile {
@@ -10,7 +10,7 @@ export class MonorepoGitIgnore implements IMetaFile {
     this.init = this.init.bind(this);
   }
 
-  public init(force = false) {
+  public init(envConfig: IEnvConfig, force = false) {
     const dest = path.resolve(process.cwd(), ".gitignore");
     const hasFile = fs.existsSync(dest);
     const gitignore = ["registry_credentials.json"];