+++ src/webpack/webpack.config.dll.ts
@@ -1,17 +1,20 @@
 import * as path from "path";
-import { rearguardConfig } from "../config/rearguard";
 import { DLL_BUNDLE_DIR_NAME } from "../const";
 import { dll_entry_name, dll_output_path, get_context } from "../helpers";
+import { IEnvConfig } from "../interfaces/config/IEnvConfig";
+import { IRearguardConfig } from "../interfaces/config/IRearguardConfig";
 import { analyze, assetsPlugin, clean, DllPlugin } from "./components/js.plugins";
 import tsLoader from "./components/ts.loaders";
 import { general_WP_config } from "./webpack.config.common";
 
 // tslint:disable:object-literal-sort-keys
 
-export function dll_WP_config() {
+export function dll_WP_config(envConfig: IEnvConfig, rearguardConfig: IRearguardConfig) {
   const { dll_entry, bundle_public_path } = rearguardConfig;
 
   return general_WP_config(
+    envConfig,
+    rearguardConfig,
     {
       [dll_entry_name()]: [path.resolve(get_context(), dll_entry)],
     },
@@ -26,10 +29,10 @@ export function dll_WP_config() {
     tsLoader(),
     [
       // ...DllReferencePlugin(true),
-      ...DllPlugin(),
-      ...assetsPlugin(DLL_BUNDLE_DIR_NAME),
-      ...analyze(),
-      ...clean([dll_output_path()]),
+      ...DllPlugin(envConfig),
+      ...assetsPlugin(envConfig, DLL_BUNDLE_DIR_NAME),
+      ...analyze(envConfig),
+      ...clean(envConfig, [dll_output_path()]),
     ],
     {},
   );