+++ CHANGELOG.md
@@ -8,6 +8,19 @@ IMPROVEMENTS:
 
 BUG FIXES:
 
+## 4.4.0 (2019-02-28)
+
+BREAKING CHANGES:
+
+- Change all --project flags to --aplication
+- Rename (package.json).rearguard.has_project to (package.json).rearguard.is_application;
+
+FEATURES:
+
+IMPROVEMENTS:
+
+BUG FIXES:
+
 ## 4.3.4 (2019-02-20)
 
 BREAKING CHANGES: